Clean code and files
