+++ server/server/common/extensions.py
@@ -1,5 +1,5 @@
-from flask_caching import Cache
 from flask import request
+from flask_caching import Cache
 
 cache = Cache(config={'CACHE_TYPE': 'simple'})
