+++ server/server/data_loader/images_files.py
@@ -1,6 +1,6 @@
-import requests
 import re
 
+import requests
 from tqdm import tqdm
