--- server/server/data_loader/segmented_data.py
@@ -1,22 +1,0 @@
-def load_segmented_data(db, segmented_data_dir):
-    docs = []
-    files = []
-    for file in segmented_data_dir.glob('**/*.json'):        
-        if file.name.startswith('_'):
-            continue
-        if '_' not in file.name:
-            continue
-        files.append(file)
-        
-    for file in files:
-        uid, muids = file.stem.split('_')
-        docs.append({
-            '_key': file.stem,
-            'uid': uid,
-            'muids': muids.split('-'),
-            'filepath': str(file.relative_to(segmented_data_dir))
-        })
-    
-    print(f'{len(docs)} files added or updated')
-    db['segmented_data'].truncate()
-    db['segmented_data'].import_bulk_logged(docs)
